requirements.txt
Adafruit-Blinka==5.4.0
adafruit-circuitpython-amg88xx==1.2.4
adafruit-circuitpython-busdevice==5.0.1
adafruit-circuitpython-register==1.9.0
Adafruit-PlatformDetect==2.17.0
Adafruit-PureIO==1.1.5
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
colour==0.1.5
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dlib==19.21.0
emoji==0.6.0
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
mpmath==1.1.0
nbclient==0.5.0
nbconvert==6.0.2
nbformat==5.0.7
nest-asyncio==1.4.0
nose==1.3.7
notebook==6.1.4
numpy==1.19.2
opencv-contrib-python==4.1.0.25
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
picamera==1.13
pickleshare==0.7.5
pkg-resources==0.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
psycopg2==2.8.6
ptyprocess==0.6.0
pycparser==2.20
pyftdi==0.51.2
Pygments==2.7.0
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
qimage2ndarray==1.8.3
qtconsole==4.7.7
QtPy==1.9.0
rpi-ws281x==4.2.4
RPi.GPIO==0.7.0
scikit-learn==0.23.2
scipy==1.5.2
scipy-stack==0.0.5
Send2Trash==1.5.0
six==1.15.0
smbus2==0.3.0
sympy==1.6.2
sysv-ipc==1.0.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==5.0.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0

